Minor Updates
+ added AdminController for app control
+ HomeController adjustments
+ RaidController added/edit dock blocks
+ added routes for the admin panel
+ created CheckAccessLevel middleware
+ raids migration added dock block
+ trainers migration new field 'verified_at'
+ added 1 second delay to the RaidCountdown for the API calls
+ added a style.scss file for custom styling
+ created views for the admin panel

+++ app/Raid.php
@@ -12,14 +12,21 @@ class Raid extends Model
 {
     protected $fillable = ['trainer_id', 'name', 'tier', 'invites', 'hatched', 'weather_boost', 'end_time', 'seconds'];
 
-    public static function safeDelete(Raid $raid)
+    /**
+     * Delete a raid after checking it exists otherwise return 'false'.
+     *
+     * @param int $id
+     * @return bool
+     */
+    public static function safeDelete(int $id): bool
     {
-        try {
-            $raid->delete();
-            return "Success!";
-        } catch (\Exception $ex) {
-            return $ex;
+        $raid = static::find($id);
+        if ($raid !== null){
+            return $raid->delete();
+        } else {
+            return false;
         }
+//        return ddd(static::findOrFail($id)->delete());
     }
 
     /**
